python/requirements.txt
bufbuild-protovalidate-community-nipunn1313-mypy==3.6.0.1.20240920164238+5a7b106cbb87
bufbuild-protovalidate-community-nipunn1313-mypy-grpc==3.6.0.1.20240920164238+5a7b106cbb87
bufbuild-protovalidate-grpc-python==1.67.0.1.20240920164238+5a7b106cbb87
bufbuild-protovalidate-protocolbuffers-pyi==28.2.0.1.20240920164238+5a7b106cbb87
bufbuild-protovalidate-protocolbuffers-python==28.2.0.1.20240920164238+5a7b106cbb87
certifi==2024.8.30
charset-normalizer==3.4.0
nativeconnect-api-community-nipunn1313-mypy==3.6.0.1.dev+1f3fb8c687d6
nativeconnect-api-community-nipunn1313-mypy-grpc==3.6.0.1.dev+1f3fb8c687d6
nativeconnect-api-grpc-python==1.67.0.1.dev+1f3fb8c687d6
nativeconnect-api-protocolbuffers-python==28.2.0.1.dev+1f3fb8c687d6
nativeconnect-api-protocolbuffers-pyi==28.2.0.1.dev+1f3fb8c687d6
compassiot-model-community-nipunn1313-mypy==3.6.0.1.dev0+4fa180fac270
compassiot-model-community-nipunn1313-mypy-grpc==3.6.0.1.20241010053502+53b3bfb524bc
compassiot-model-grpc-python==1.67.0.1.20241010053502+53b3bfb524bc
compassiot-model-protocolbuffers-pyi==28.2.0.1.20241010053502+53b3bfb524bc
compassiot-model-protocolbuffers-python==28.2.0.1.20241010053502+53b3bfb524bc
googleapis-googleapis-grpc-python==1.67.0.1.20240827201746+e7f8d366f526
googleapis-googleapis-protocolbuffers-pyi==28.2.0.1.20240827201746+e7f8d366f526
googleapis-googleapis-protocolbuffers-python==28.2.0.1.20240827201746+e7f8d366f526
googleapis_googleapis_community_nipunn1313_mypy==3.6.0.1.20240827201746+e7f8d366f526
googleapis_googleapis_community_nipunn1313_mypy_grpc==3.6.0.1.20240827201746+e7f8d366f526
grpc-interceptor==0.15.4
grpc-stubs==1.53.0.5
grpcio==1.66.2
grpcio-tools==1.66.2
idna==3.10
protobuf==5.28.2
requests==2.32.3
setuptools==75.1.0
types-protobuf==5.28.0.20240924
urllib3==2.2.3
